Unify NorthstarLauncher and wsock32 proxy
