+++ primedev/CMakeLists.txt
@@ -1,3 +1,2 @@
 include(Northstar.cmake)
 include(Launcher.cmake)
-include(WSockProxy.cmake)
